requirements-dev.txt
# Testing Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Development Tools
pre-commit>=3.3.3
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
bandit>=1.7.5

# Type Stubs
types-PyYAML>=6.0.12
pandas-stubs>=2.0.3 